refactor: rename main components package
