+++ socceranalyzer/common/analysis/ball_possession.py
@@ -3,7 +3,7 @@ import pandas as pd
 from socceranalyzer.common.enums.sim2d import SIM2D
 from socceranalyzer.common.enums.ssl import SSL
 from socceranalyzer.utils.logger import Logger
-from socceranalyzer.common.chore.mediator import Mediator
+from socceranalyzer.common.core.mediator import Mediator
 from socceranalyzer.common.geometric.point import Point
 from socceranalyzer.common.operations.measures import distance_sqrd
