+++ socceranalyzer/common/analysis/goalkeeper.py
@@ -5,7 +5,7 @@ from socceranalyzer.common.enums.ssl import SSL
 from socceranalyzer.common.enums.vss import VSS
 from socceranalyzer.common.analysis.find_goals import FindGoals
 from socceranalyzer.common.geometric.point import Point
-from socceranalyzer.common.chore.mediator import Mediator
+from socceranalyzer.common.core.mediator import Mediator
 from socceranalyzer.common.operations.measures import distance
 from socceranalyzer.utils.logger import Logger
