+++ socceranalyzer/common/analysis/heatmap.py
@@ -5,7 +5,7 @@ import os
 
 from socceranalyzer.common.basic.match import Match
 from socceranalyzer.common.analysis.abstract_analysis import AbstractAnalysis
-from socceranalyzer.common.chore.mediator import Mediator
+from socceranalyzer.common.core.mediator import Mediator
 from socceranalyzer.common.enums.sim2d import SIM2D
 from socceranalyzer.common.enums.ssl import SSL
 from socceranalyzer.common.enums.vss import VSS
