+++ socceranalyzer/common/analysis/stamina.py
@@ -1,7 +1,7 @@
 import pandas
 from socceranalyzer.common.analysis.abstract_analysis import AbstractAnalysis
 from socceranalyzer.common.enums.sim2d import SIM2D
-from socceranalyzer.common.chore.mediator import Mediator
+from socceranalyzer.common.core.mediator import Mediator
 from socceranalyzer.common.enums.ssl import SSL
 from socceranalyzer.common.enums.vss import VSS
 from socceranalyzer.utils.logger import Logger
