+++ socceranalyzer/common/core/abstract_factory.py
