+++ socceranalyzer/common/evaluators/ball_holder.py
@@ -6,7 +6,7 @@ from socceranalyzer.common.enums.vss import VSS
 from socceranalyzer.common.geometric.point import Point
 
 from socceranalyzer.common.geometric.circle import Circle
-from socceranalyzer.common.chore.mediator import Mediator
+from socceranalyzer.common.core.mediator import Mediator
 from socceranalyzer.common.operations.measures import distance
 
 class BallHolderEvaluator:
