+++ socceranalyzer/common/evaluators/passing.py
@@ -1,6 +1,6 @@
 from typing import Any
 from pandas import DataFrame
-from socceranalyzer.common.chore.mediator import Mediator
+from socceranalyzer.common.core.mediator import Mediator
 from socceranalyzer.common.enums.sim2d import SIM2D
 from socceranalyzer.common.enums.ssl import SSL
 from socceranalyzer.common.enums.vss import VSS
